update the fork

+++ src/test/mbtest1.cpp
@@ -13,7 +13,6 @@ using namespace mabain;
 
 static std::atomic<int64_t> key_low;
 static std::atomic<int64_t> key_high;
-static int64_t key_diff = 3000000;
 static int64_t memcap_i = 256*1024*1024;
 static int64_t memcap_d = 256*1024*1024;
 static pthread_t wid = 0;
@@ -42,22 +41,22 @@ static void* Reader(void *arg)
     int ktype;
 
     while(!stop_processing) {
-        usleep(5);
+        //usleep(5);
 
 	if(key_high.load(std::memory_order_consume) == key_low.load(std::memory_order_consume))
 	    continue;
-        ikey = key_low.load(std::memory_order_consume); 
+        ikey = key_low.load(std::memory_order_consume);
         ikey += rand() % (key_high.load(std::memory_order_consume) - key_low.load(std::memory_order_consume));
- 
+
         ktype = rand() % 3;
         switch(ktype) {
-        case 0: 
+        case 0:
             keystr = tkey_int.get_key(ikey);
             break;
-        case 1: 
+        case 1:
             keystr = tkey_sha1.get_key(ikey);
             break;
-        case 2: 
+        case 2:
             keystr = tkey_sha2.get_key(ikey);
             break;
         }
@@ -75,7 +74,7 @@ static void* Reader(void *arg)
             std::cout << "unexpected return from Find: " << rval << "\n";
             abort();
         }
-    }    
+    }
 
     db.Close();
     return NULL;
@@ -117,7 +116,7 @@ static void Verify(DB &db)
 static void* AddThread(void *arg)
 {
     DB *db = (DB *) arg;
-    int num = rand() % 50;
+    int num = rand() % 1500;
 
     int64_t key;
     TestKey tkey_int(MABAIN_TEST_KEY_TYPE_INT);
@@ -140,7 +139,7 @@ static void* AddThread(void *arg)
 }
 
 static void Populate(DB &db, int nt)
-{   
+{
     pthread_t tid[256];
     assert(nt < 256);
     for(int i = 0; i < nt; i++) {
@@ -149,7 +148,7 @@ static void Populate(DB &db, int nt)
             abort();
         }
     }
-    
+
     for(int i = 0; i < nt; i++) {
         if(pthread_join(tid[i], NULL) != 0) {
             std::cout << "failed to join MultiThreadAdd thread\n";
@@ -179,18 +178,19 @@ static void load_key_ids()
 
 static void store_key_ids()
 {
-    std::ofstream ofs;
     std::string path = mbdir + "/key_id";
-    ofs.open (mbdir.c_str(), std::ofstream::out);
-    ofs << key_low << "\n";
-    ofs << key_high << "\n";
-    ofs.close();
+    FILE *fp;
+    fp = fopen(path.c_str(), "w");
+    if(!fp) return;
+    fprintf(fp, "%d\n%d", (int)key_low.load(std::memory_order_consume),
+                          (int)key_high.load(std::memory_order_consume));
+    fclose(fp);
 }
 
 static void* DeleteThread(void *arg)
 {
     DB *db = (DB *) arg;
-    int num = rand() % 10000;
+    int num = rand() % 5;
 
     int64_t key;
     TestKey tkey_int(MABAIN_TEST_KEY_TYPE_INT);
@@ -238,8 +238,20 @@ void stop_mb_test()
     if(wid != 0) {
         if(pthread_join(wid, NULL) != 0) {
             std::cout << "cannot join mbtest thread\n";
-	}
+	      }
     }
+}
+
+static void CheckCount(DB *db)
+{
+    if(db == NULL) return;
+
+    int64_t count = 0;
+    for(DB::iterator iter = db->begin(false, false); iter != db->end(); ++iter) {
+        count++;
+    }
+    std::cout << "Count using iterator: " << count << "\tcount from API: "
+              << db->Count() << "\n";
 }
 
 void start_mb_test()
@@ -258,8 +270,20 @@ static void* run_mb_test(void *arg)
 
     load_key_ids();
 
-    int options = CONSTS::WriterOptions() | CONSTS::ASYNC_WRITER_MODE;
-    DB *db = new DB(mbdir.c_str(), options, memcap_i, memcap_d);
+    MBConfig mbconf;
+    int options;
+    memset(&mbconf, 0, sizeof(mbconf));
+    mbconf.mbdir = mbdir.c_str();
+    options = CONSTS::WriterOptions() | CONSTS::ASYNC_WRITER_MODE;
+    mbconf.options = options;
+    mbconf.memcap_index = 128ULL*1024*1024;
+    mbconf.memcap_data = 128ULL*1024*1024;
+    mbconf.block_size_index = 64U*1024*1024;
+    mbconf.block_size_data = 64U*1024*1024;
+    mbconf.max_num_data_block = 3;
+    mbconf.max_num_index_block = 3;
+    mbconf.num_entry_per_bucket = 500;
+    DB *db = new DB(mbconf);
     if(!db->is_open()) {
         std::cerr << "failed to open writer db" << std::endl;
 	delete db;
@@ -275,44 +299,42 @@ static void* run_mb_test(void *arg)
     }
 
     int rcn = 0;
+    int64_t loop_cnt = 0;
     while(!stop_processing) {
-        if(key_high - key_low < key_diff) {
-            Populate(*db, nupdates);
-        } else {
-            if(db->Count() < key_diff*3 - 1000)
-                continue;
+        Populate(*db, nupdates);
+        Prune(*db, nupdates);
 
-            std::cout << key_low.load(std::memory_order_consume) << ": "
-                      << key_high.load(std::memory_order_consume) << std::endl;
-
-            Prune(*db, nupdates);
-
-            int rval = db->CollectResource(10LL*1024*1024, 10LL*1024*1024, 0xFFFFFFFFFFFF, 0xFFFFFFFFFFFF);
+        std::cout << "LOOP " << loop_cnt << ": " << db->Count() << "\n";
+        if(loop_cnt % 5 == 0) {
+            std::cout << "RUN RC\n";
+            int rval = db->CollectResource(24LL*1024*1024, 24LL*1024*1024, 128LL*1024*1024, 0xFFFFFFFFFFFF);
             if(rval == MBError::SUCCESS) {
                 rcn++;
                 if(rcn % 57 == 0 && !(options & CONSTS::ASYNC_WRITER_MODE)) {
                     // Note if in async mode, the DB handle cannot be used for lookup.
                     std::cout << "Verifying after rc" << std::endl;
                     Verify(*db);
-                }
-		else if(rcn % 100 == 0) {
-		    db->Close();
-		    delete db;
-		    db = new DB(mbdir.c_str(), options, memcap_i, memcap_d);
-		    if(!db->is_open()) {
+                } else if(rcn % 20 == 0) {
+                    db->Close();
+	            delete db;
+                    mbconf.options = options;
+	            db = new DB(mbconf);
+	            if(!db->is_open()) {
 	                delete db;
-			abort();
-		    }
-		}
-                else if(rcn % 20000523 == 0) {
+	                abort();
+	            }
+                    CheckCount(db);
+                } else if(rcn % 20000523 == 0) {
                     if(system("reboot") != 0) {
                     }
                 }
             }
         }
 
+        sleep(1);
+        loop_cnt++;
         if(time(NULL) >= run_stop_time) {
-            stop_processing = true;   
+            stop_processing = true;
         }
     }
 
@@ -320,7 +342,7 @@ static void* run_mb_test(void *arg)
     delete db;
     store_key_ids();
 
-    
+
     return NULL;
 }
